requirements.txt
flet>=0.21.0
httpx>=0.25.0
python-dotenv>=1.0.0